Cooked 5 star meal
We need to test values for the forces and distances. We need to turn the boids into triangles as well I'd say. If we want to play bad apple there's an algorithm we can leverage for that but we'll need to look into it. If not we can just color them somehow, maybe add a predator or some shit.